Rename getAdminMetaSchema.ts to adminMetaSchema.ts
