+++ packages/core/src/admin-ui/system/index.ts
@@ -1,3 +1,3 @@
 export { generateAdminUI } from './generateAdminUI';
-export { KeystoneMeta } from './getAdminMetaSchema';
+export { KeystoneMeta } from './adminMetaSchema';
 export { buildAdminUI } from './buildAdminUI';
